Change default schema when creating/updating package.
We add some fields to the pkg_dict as lists of dicts, which get added to the '__junk' field on package creation/updating. This then causes a validation error because the default package creation schema validates that '__junk' is empty.

In the Ogdch plugin in ckanext-switzerland, we include a validator that moves these fields into the package dict properly (list_of_dics).